requirements.txt
streamlit
python-dotenv
requests
# openai==1.54.5  # 已改为HTTP直连，不再强制需要，可选
